review resolved - testcases

+++ backend/controllers/blogController.js
@@ -1,13 +1,14 @@
 const blogService = require("../service/blogService");
 const { sendResponse } = require("../utils/contentNegotiation");
-const { blogValidation } = require("../utils/blogValidation");
+const { blogValidation} = require("../utils/blogValidation");
 const { paginate } = require("../utils/pagination");
 
 ("use strict");
 
 exports.getAllBlogs = async (req, res, next) => {
   try {
-    const { offset, limit } = paginate(req);
+    
+    const { offset, limit } = paginate(req.query.pageNo, req.query.pageSize);
     const allBlogs = await blogService.getAllBlogs(offset, limit);
     return sendResponse(req, res, allBlogs.status, allBlogs.message);
   } catch (error) {
@@ -31,7 +32,7 @@ exports.createBlog = async (req, res, next) => {
 
     blogValidation(title, description);
 
-    const createdBlog = await blogService.createBlog(req.body);
+    const createdBlog = await blogService.createBlog(title, description);
     return sendResponse(req, res, createdBlog.status, createdBlog.message);
   } catch (error) {
     next(error);
@@ -44,7 +45,9 @@ exports.updateBlog = async (req, res, next) => {
     const title = req.body.title;
     const description = req.body.description;
 
-    blogValidation(title, description);
+    if(!blogId) throw Object.assign(new Error("Enter a valid blogId!"), { statusCode: 400 });
+    
+    blogValidation( title, description);
 
     const updatedBlog = await blogService.updateBlog(
       blogId,
@@ -60,6 +63,8 @@ exports.updateBlog = async (req, res, next) => {
 exports.deleteBlog = async (req, res, next) => {
   try {
     const blogId = req.params.blogId;
+    if(!blogId) throw Object.assign(new Error("Enter a valid blogId!"), { statusCode: 400 });
+
     const deletedBlog = await blogService.deleteBlog(blogId);
     return sendResponse(req, res, deletedBlog.status, deletedBlog.message);
   } catch (error) {
